requirements.txt
google-api-python-client
python-dotenv
requests
yt-dlp
sqlalchemy
pymysql
panads